build: updated deps for improved webpack output
