+++ demo/vanilla/views/index.pug
@@ -113,7 +113,7 @@ html(lang="en")
         script(src="/tsparticles-basic/tsparticles.basic.min.js")
         script(src="/tsparticles-pjs/tsparticles.pjs.min.js")
         script(src="/tsparticles-slim/tsparticles.slim.min.js")
-        script(src="/tsparticles/tsparticles.js")
+        script(src="/tsparticles/tsparticles.min.js")
         script(src="/tsparticles-all/tsparticles.all.min.js")
         script(src="/interaction-external-attract/tsparticles.interaction.external.attract.min.js")
         script(src="/interaction-external-bounce/tsparticles.interaction.external.bounce.min.js")
