+++ package.json
@@ -25,32 +25,32 @@
   },
   "prettier": "@tsparticles/prettier-config",
   "devDependencies": {
-    "@babel/core": "^7.23.7",
-    "@commitlint/cli": "^18.4.4",
-    "@commitlint/config-conventional": "^18.4.4",
+    "@babel/core": "^7.23.9",
+    "@commitlint/cli": "^18.6.0",
+    "@commitlint/config-conventional": "^18.6.0",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
     "@nx/workspace": "^17.2.8",
-    "@tsparticles/cli": "^2.0.4",
-    "@tsparticles/eslint-config": "^2.0.4",
-    "@tsparticles/prettier-config": "^2.0.1",
+    "@tsparticles/cli": "^2.0.5",
+    "@tsparticles/eslint-config": "^2.1.0",
+    "@tsparticles/prettier-config": "^2.1.0",
     "@tsparticles/tsconfig": "^2.0.1",
-    "@tsparticles/webpack-plugin": "^2.0.4",
+    "@tsparticles/webpack-plugin": "^2.1.3",
     "@types/chai": "^4.3.11",
     "@types/jsdom": "^21.1.6",
     "@types/mocha": "^10.0.6",
-    "@types/node": "^20.10.7",
+    "@types/node": "^20.11.9",
     "@types/webpack-env": "^1.18.4",
-    "@typescript-eslint/eslint-plugin": "^6.18.0",
-    "@typescript-eslint/parser": "^6.18.0",
+    "@typescript-eslint/eslint-plugin": "^6.19.1",
+    "@typescript-eslint/parser": "^6.19.1",
     "babel-loader": "^9.1.3",
-    "browserslist": "^4.22.2",
+    "browserslist": "^4.22.3",
     "canvas": "^2.11.2",
     "chai": "^4.4.0",
     "copyfiles": "^2.4.1",
     "cross-env": "^7.0.3",
     "eslint": "^8.56.0",
     "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-jsdoc": "^48.0.2",
+    "eslint-plugin-jsdoc": "^48.0.4",
     "eslint-plugin-tsdoc": "^0.2.17",
     "gh-pages": "^6.1.1",
     "husky": "^9.0.6",
@@ -61,7 +61,7 @@
     "nx": "^17.2.8",
     "nx-cloud": "^16.5.2",
     "nyc": "^15.1.0",
-    "prettier": "^3.1.1",
+    "prettier": "^3.2.4",
     "prettier-plugin-multiline-arrays": "^3.0.1",
     "rimraf": "^5.0.5",
     "source-map-support": "^0.5.21",
@@ -73,11 +73,11 @@
     "typedoc-plugin-carbon-ads": "^1.6.0",
     "typedoc-plugin-coverage": "^2.2.0",
     "typedoc-plugin-keywords": "^1.6.0",
-    "typedoc-plugin-mdn-links": "^3.1.11",
-    "typedoc-plugin-missing-exports": "^2.1.0",
+    "typedoc-plugin-mdn-links": "^3.1.14",
+    "typedoc-plugin-missing-exports": "^2.2.0",
     "typescript": "^5.3.3",
     "typescript-json-schema": "^0.62.0",
-    "webpack": "^5.89.0",
+    "webpack": "^5.90.0",
     "webpack-bundle-analyzer": "^4.10.1",
     "webpack-cli": "^5.1.4",
     "yargs": "^17.7.2"
