+++ shapes/emoji/src/Constants.ts
@@ -1,0 +1,1 @@
+export const validTypes = ["emoji"];
