+++ shapes/image/src/index.ts
@@ -1,4 +1,4 @@
-import { type IImage, downloadSvgImage, loadGifImage, loadImage } from "./Utils.js";
+import { type IImage, downloadSvgImage, loadImage } from "./Utils.js";
 import type { IPreload } from "./Options/Interfaces/IPreload.js";
 import type { ImageEngine } from "./types.js";
 import { errorPrefix } from "@tsparticles/engine";
@@ -44,6 +44,8 @@ function addLoadImageToEngine(engine: ImageEngine): void {
             let imageFunc: (image: IImage) => Promise<void>;
 
             if (data.gif) {
+                const { loadGifImage } = await import("./GifUtils/Utils.js");
+
                 imageFunc = loadGifImage;
             } else {
                 imageFunc = data.replaceColor ? downloadSvgImage : loadImage;
