+++ updaters/outModes/src/BounceOutMode.ts
@@ -6,7 +6,6 @@ import {
     type Particle,
     calculateBounds,
 } from "@tsparticles/engine";
-import { bounceHorizontal, bounceVertical } from "./Utils.js";
 import type { IOutModeManager } from "./IOutModeManager.js";
 
 export class BounceOutMode implements IOutModeManager {
@@ -19,12 +18,12 @@ export class BounceOutMode implements IOutModeManager {
         ];
     }
 
-    update(
+    async update(
         particle: Particle,
         direction: OutModeDirection,
         delta: IDelta,
         outMode: OutMode | keyof typeof OutMode,
-    ): void {
+    ): Promise<void> {
         if (!this.modes.includes(outMode)) {
             return;
         }
@@ -34,7 +33,7 @@ export class BounceOutMode implements IOutModeManager {
 
         for (const [, plugin] of container.plugins) {
             if (plugin.particleBounce !== undefined) {
-                handled = plugin.particleBounce(particle, delta, direction);
+                handled = await plugin.particleBounce(particle, delta, direction);
             }
 
             if (handled) {
@@ -50,7 +49,8 @@ export class BounceOutMode implements IOutModeManager {
             offset = particle.offset,
             size = particle.getRadius(),
             bounds = calculateBounds(pos, size),
-            canvasSize = container.canvas.size;
+            canvasSize = container.canvas.size,
+            { bounceHorizontal, bounceVertical } = await import("./Utils.js");
 
         bounceHorizontal({ particle, outMode, direction, bounds, canvasSize, offset, size });
         bounceVertical({ particle, outMode, direction, bounds, canvasSize, offset, size });
